fix: avoid out-of-memory on large datasets
Do not asynchronously queue writes of the output files when all the
renders are happening on the same the same tick. On large datasets,
this causes the output to stay in memory until all the renders are done.

In a pahologic case we observed heap usage grow to 22GB. After the fix
it remains around 1GB.